implemeneted GUID instead of string ID, and RESTful elements (returns 404 properly now)

+++ Controllers/ItemController.cs
@@ -1,9 +1,11 @@
 ﻿using Cogburn_Shop.Entities;
 using Cogburn_Shop.Services;
+using Cogburn_Shop.DTOs;
 using Microsoft.AspNetCore.Mvc;
 using MongoDB.Driver;
 using MongoDB.Bson;
 using Microsoft.Extensions.Options;
+using ZstdSharp.Unsafe;
 
 namespace Cogburn_Shop.Controllers
 {
@@ -22,12 +24,13 @@ namespace Cogburn_Shop.Controllers
         [HttpGet]
         public async Task<List<Item>> GetItems()
         {
-            return await _repository.GetAsync();
+            return await _repository.GetAsync(); 
         }
-                
+          
+        
         // GET /items/{id}
         [HttpGet("{id}")]
-        public async Task<ActionResult<Item>> Get(string id)
+        public async Task<ActionResult<Item>> Get(Guid id)
         {
             var item = await _repository.GetAsync(id);
 
@@ -39,27 +42,54 @@ namespace Cogburn_Shop.Controllers
             return item;
         }
         
+
         //POST /item
         [HttpPost]
-        public async Task<IActionResult> Post([FromBody] Item item)
+        public async Task<IActionResult> Post(Item item)
         {
+            item = new()
+            {
+                Id = Guid.NewGuid(),
+                Name = item.Name,
+                Price = item.Price,
+                Description = item.Description
+            };
+            
             await _repository.CreateAsync(item);
-            return CreatedAtAction(nameof(GetItems), new { item = item.Id }, item);
+            return CreatedAtAction(nameof(GetItems), new { id = item.Id }, item);
         }
+
 
         //PUT /item/{id}
         [HttpPut("{id}")]
-        public async Task<IActionResult> UpdateAsync(string id, [FromBody] Item updatedItem)
+        public async Task<IActionResult> UpdateAsync(Guid id, [FromBody] Item updatedItem)
         {
+            var item = await _repository.GetAsync(id);
+
+            if (item == null)
+            {
+                return NotFound();
+            }
+
             await _repository.UpdateAsync(id, updatedItem);
+
             return NoContent();
         }
 
+
         //DELETE /item/{id}
         [HttpDelete("{id}")]
-        public async Task<IActionResult> DeleteAsync(string id)
+        public async Task<IActionResult> DeleteAsync(Guid id)
         {
+            var item = await _repository.GetAsync(id);
+
+            if (item == null)
+            {
+                return NotFound();
+            }
+
             await _repository.DeleteAsync(id);
+
             return NoContent();
         }
     }
